+++ crates/vm/src/arch/execution_mode/metered/segment_ctx.rs
@@ -40,7 +40,7 @@ impl Default for SegmentationLimits {
 pub struct SegmentationCtx {
     pub segments: Vec<Segment>,
     pub(crate) air_names: Vec<String>,
-    widths: Vec<usize>,
+    pub(crate) widths: Vec<usize>,
     interactions: Vec<usize>,
     pub(crate) segmentation_limits: SegmentationLimits,
     pub instret_last_segment_check: u64,
@@ -152,13 +152,13 @@ impl SegmentationCtx {
             return false;
         }
 
-        for (i, (height, is_constant)) in trace_heights
+        for (i, (&height, is_constant)) in trace_heights
             .iter()
             .zip(is_trace_height_constant.iter())
             .enumerate()
         {
             // Only segment if the height is not constant and exceeds the maximum height
-            if !is_constant && *height > self.segmentation_limits.max_trace_height {
+            if !is_constant && height > self.segmentation_limits.max_trace_height {
                 let air_name = &self.air_names[i];
                 tracing::info!(
                     "Segment {:2} | instret {:9} | chip {} ({}) height ({:8}) > max ({:8})",