+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to OpenVM will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project follows a versioning principles documented in [VERSIONING.md](./VERSIONING.md).
 
+## Unreleased
+
+### Changed
+- (Executor) Modified `VirtualMachine::build_metered_ctx` to take the program (`&VmExe<Val<E::SC>>`) as an argument.
+
 ## v1.4.0 (2025-09-01)
 
 ### Added